Added some wacky images because for some reason I was getting errors when trying to reference local assets?
- So I had to do some very rudimentary p5 drawing, and link to gifs online lol

Just FYI I had some issues during this exercise and therefore wasn't able to make it as fleshed out as I would have liked.

I had a lot of trouble referencing specific objects in different JSON files, even though following along with Pippin's structure seems very straightforward.

For example:
https://github.com/dariusk/corpora/blob/master/data/humans/richpeople.json

I wasn't able to reference the "name" object properly and I'm not sure what I was doing wrong. I'd figure it would be something like "object.richPeople[0].name" but that didn't work. So I must be doing something wrong!  I ran into similar issues with other JSON files that had very different formats which was frustrating (mostly because I had a concept that I thought would be fun--to make the user some kind of horrible corporate assassin).

Also, for some reason when I loaded assets (JSON or images), I would get an error in the console when one of the assets was a link to an online source and one was stored in my local assets folder. Not sure why this was an issue!

Anywho, thanks for your time and for the extension!