Added a new function to AudioObject that fades a gif in and out over top of the other images. However, it is not working properly!
Can't seem to figure out how to make the tint image in the darknessImgDisplay function only affect the darknessImg and not the other images as well.

I tried using push() pop() but then it stops the fade from working at all.

I also tried placing the other images above the audioObject.darknessImgDisplay call, but this results in them not appearing at all.